Open AI App To Summarize Text Data/requirements.txt
Flask==2.1.3
openai==0.26.5
python-dotenv==1.0.0
requests==2.28.1
gunicorn==20.1.0